# doc/api.yaml
openapi: 3.0.3
info:
  title: WASAPhoto
  description: |-
    Keep in touch with your friends by sharing photos of special moments, thanks to WASAPhoto! 
    You can upload your photo directly from your PC, and they will be visible to everyone following you
  version: 1.0.0

#============================================TAGS PART==============================================#
...
security:
  - bearerAuth: []
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      description: Bearer token for the WASAPhoto API.
      scheme: bearer

#============================================SCHEMAS PART==============================================#
  schemas:
...
    Code-Identifier:
      title: Code-random
      type: object
      description: |-
        General random generated string, which will be used as a unique identifier.
        It is a unique code that identifier uniquely:
          - the user
          - the photo
          - the comment
      properties:
        identifier: 
          type: string
          description: unique identifier generatedr randomly
          pattern: '^.*?$'
          minLength: 64
          maxLength: 64
          example: "115981bedf5cc6956330a1e4c394526a8839e8a3ec3adf0fre3eedd59c759a58"
      
    
#_____________________________USER FIELD____________________________________________
#-------------------------------------------------------------------------------- 
    Username:
      title: Username
      description: |-
        The username is an invented name that user assign.
        It can be composed by:
         - number 
         - alphabetical character
         - dot
         - underscore
        The length of the username is minimum 6 character and maximum 12 character 
      type: object
      properties:
        username-string: 
          type: string
          description: username string 
          minLength: 6
          maxLength: 12
          pattern: '^[a-zA-Z0-9_.]{6,12}$'
          example: "leo.ramill"

      example:
        username-string: "leo.ramill"
#--------------------------------------------------------------------------------
    UserID:
      title: UserID
      description: It's an object that contains an identifier code of user
      type: object
      properties:
        code-user: 
          $ref: "#/components/schemas/Code-Identifier"
      
#--------------------------------------------------------------------------------
    UserProfile:
      title: User Profile
      description: |-
        For every user there are an information about:
          - user id 
          - nickname
          - followers
          - ollowing
          - posts
      type: object
      properties:
        user-identifier:
           $ref: "#/components/schemas/Code-Identifier/properties/identifier"
        nickname:
          $ref: "#/components/schemas/Username/properties/username-string"
        followers:
          type: array
          description: List of follower
          minItems: 1
          maxItems: 1000
          items:
            $ref: "#/components/schemas/CompleteUser"
        following:
          type: array
          description: List of following
          minItems: 1
          maxItems: 1000
          items:
            $ref: "#/components/schemas/CompleteUser"
        posts:
          type: array
          description: List of PostedPhoto
          minItems: 1
          maxItems: 1000
          items:
            $ref: "#/components/schemas/PostedPhoto"

#--------------------------------------------------------------------------------        
    CompleteUser:
      title: Complete User
      description: |-
        For every user there are a combination of username and userID
      type: object
      properties:
        user-id:
           $ref: "#/components/schemas/UserID"
        username:
          $ref: "#/components/schemas/Username"


#--------------------------------------------------------------------------------    
...
    CommentID:
      title: Comment ID
      description: It's a object that contains an identifier code of comment
      type: object
      properties:
        code-comment:
          $ref: "#/components/schemas/Code-Identifier"
        
#--------------------------------------------------------------------------------  
    Comment:
      title: Comment
      description: |-
        Comment is an object which are strings writing by Users. It's composed by:
          - comment id
          - date and time that he/she posted the comment
          - the body string of the comment
          - owner comment
      type: object
      properties:
        comment-id:
          $ref: "#/components/schemas/CommentID"

        photo-id: 
          $ref: "#/components/schemas/ImageID"
        
        date-time:
          type: string
          description: Date and Time that was published the comment
          format: date-time
          pattern: ^\d{4}-\d{2}-\d{2}T\d{2}:\d{2}:\d{2}Z$
          minLength: 20
          maxLength: 20
          readOnly: true
        
        text:
          type: string
          description: body string of the comment
          pattern: ^[^\\]{1,1024}$
          minLength: 1
          maxLength: 1024
        
        user-id: 
          $ref: "#/components/schemas/UserID"

        from-user: 
          $ref: "#/components/schemas/Username/properties/username-string"

        
      



#_____________________________LIKE FIELD____________________________________________
#--------------------------------------------------------------------------------
    LikeID:
      title: Like ID
      description: It's a object that contains an identifier code of like
      type: object
      properties:
        code-like: 
          $ref: "#/components/schemas/Code-Identifier"
        
#-------------------------------------------------------------------------------- 
    Like:
      title: Like
      description: |-
        Like is an object which are put in the photo by Users. It is composed by:
          - Like ID 
          - Owner like
      type: object
      properties:
        like-id:
          $ref: "#/components/schemas/LikeID"
  
        photo-id:
          $ref: "#/components/schemas/ImageID"
        
        from-user: 
          $ref: "#/components/schemas/Username/properties/username-string"
        


#_____________________________PHOTO FIELD____________________________________
#----------------------------------------------------------------------------

#-------------------------------------------------------------------------------- 
...
    ImageID:
      title: ImageID
      description: It's an object that contains an identifier code of image
      type: object
      properties:
        code-image: 
          $ref: "#/components/schemas/Code-Identifier"
          
#-------------------------------------------------------------------------------- 
    PostedPhoto:
      title: PostedPhoto
      description: |-
        PostedPhoto is a photo that was posted in the WASAPhoto website and appears:
          - in your profile 
          - in the follower stream
        It's object composed by:
          - photo_id
          - username
          - image
          - description 
          - creation time
          - comment
          - like
      type: object
      properties:
        photo-id: 
          $ref: "#/components/schemas/ImageID"

        nickname: 
          $ref: "#/components/schemas/Username/properties/username-string"

        user-id:
          $ref: "#/components/schemas/UserID"
        
        imagedata: 
          type: string
          description: URL of the imgae resource
          minLength: 0
          maxLength: 500
          pattern: ^[^\\]{0,500}$

        like-collection:
          type: array
          description: List of Like
          items:
            $ref: "#/components/schemas/Like"
          minItems: 1
          maxItems: 1000

        comment-collection:
          type: array
          description: List of comment
          items:
            $ref: "#/components/schemas/Comment"
          minItems: 1
          maxItems: 1000
        
        description-post:
          type: string
          description: body of string in which the user describe the PostedPhoto
          minLength: 0
          maxLength: 1000
          pattern: ^[^\\]{0,1000}$
          
        date-time:
          type: string
          description: Date and Time that was published the comment
          format: date-time
          pattern: ^\d{4}-\d{2}-\d{2}T\d{2}:\d{2}:\d{2}Z$
          minLength: 20
          maxLength: 20
          readOnly: true
        
#-------------------------------------------------------------------------------- 
...
    Error:
      title: Error
      type: object
      description: |-
        An error message that contains:
          - Code 
          - Message
      properties:
        code-error:
          type: integer
          description: The error code
          minimum: 3
          maximum: 3

        message-error:
          type: string
          description: The error message
          minLength: 1
          maxLength: 256
          pattern: ^[^\\]{1,256}$

#============================================PARAMETERS PART==============================================#
  parameters:
    user-id:
      name: user-id
      in: path
      description: The ID of a user to get information about him/her
      required: true
      schema: 
        $ref: "#/components/schemas/Username/properties/username-string"
      
#--------------------------------------------------------------------------------
...
paths:
...
  /users/{user-id}/profile:
    parameters: 
      - $ref: "#/components/parameters/user-id"

    put:
      tags: ["user", "login"]
      summary: Modify Username
      description: |-
        User update its username after the creation in the login.
        Returns the new identifier in the response body.
      operationId: setMyUserName
      
      requestBody:
        content:
          application/json:
            schema: 
              $ref: '#/components/schemas/Username'
        required: true
      
      responses:
        "201":
          description: "The user's name has been updated"
          content:
            application/json:
              schema: 
                $ref: "#/components/schemas/Username/properties/username-string"

        "400":
          description: Invalid user-id
          content:
            application/json:
              schema:  
                $ref: "#/components/schemas/Error"

        "401":
          description: |-
            The UserInfo is invalid, no authentication information
            has been provided
          content:
            application/json:
              schema:  
                $ref: "#/components/schemas/Error"


        "500":
          description: Internal Server Error
          content:
            application/json:
              schema: 
                $ref: "#/components/schemas/Error"

    get:
      tags: [ "user"]
      summary: Get User Profile 
      description: Get Profile information
      operationId: getUserProfile
      
      responses:
        "200":
          description: "Information about Profile"
          content:
            application/json:
              schema: 
                $ref: "#/components/schemas/UserProfile"

        "400":
          description: Bad Request
          content:
            application/json:
              schema: 
                $ref: "#/components/schemas/Error"

        "401":
          description: |-
             The requested user-id is invalid, no authentication information
             has been provided
          content:
            application/json:
              schema: 
                $ref: "#/components/schemas/Error"

        "404":
          description: The user is not found
          content:
            application/json:
              schema: 
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema: 
                $ref: "#/components/schemas/Error"



#_____________________________PROFILE - BANNED___________________________
#----------------------------------------------------------------------
...